LayoutTests/media/media-source/media-source-real-abort.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-abort</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Tests that abort() properly resets parser state, allowing a new
    // init segment and media segment to be appended successfully.
    // This exercises the real SourceBufferParserAVFObjC reset path.

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        source = new MediaSource();
        video.src = URL.createObjectURL(source);
        await waitFor(source, 'sourceopen');

        sourceBuffer = source.addSourceBuffer(MP4.VIDEO_TYPE);

        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });

        // First: append init + media normally
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        var media1 = MP4.mediaSegment({
            sequenceNumber: 1,
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 0, samples: [
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
            ]}]
        });
        sourceBuffer.appendBuffer(media1);
        await waitFor(sourceBuffer, 'updateend');

        testExpected('sourceBuffer.buffered.length', 1);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.end(0)', 2, 0.01);

        // Abort resets the parser state
        run('sourceBuffer.abort()');
        testExpected('sourceBuffer.updating', false);

        // After abort, re-send init + new media segment
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        var media2 = MP4.mediaSegment({
            sequenceNumber: 2,
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 2000, samples: [
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
            ]}]
        });
        sourceBuffer.appendBuffer(media2);
        await waitFor(sourceBuffer, 'updateend');

        // Should now have [0, 4) — original 2s + new 2s contiguous
        testExpected('sourceBuffer.buffered.length', 1);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.start(0)', 0, 0.01);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.end(0)', 4, 0.01);
        testExpected('source.readyState', 'open');

        endTest();
    }
    </script>
</head>
<body onload="runTest()">
    <div>Tests abort() followed by re-initialization through the real parser pipeline.</div>
    <video></video>
</body>
</html>
